requirements.txt
fastapi >=0.77.1,<1.0
uvicorn[standard] >=0.17.6,<1.0
pynacl >=1.5.0,<2.0